+++ tests/core/test_mixer.py
@@ -23,6 +23,7 @@ class CoreMixerTest(unittest.TestCase):
         self.mixer.get_volume.assert_called_once_with()
 
     def test_set_volume(self):
+        self.mixer.set_volume.return_value.get.return_value = True
         self.core.mixer.set_volume(30)
 
         self.mixer.set_volume.assert_called_once_with(30)
@@ -34,6 +35,7 @@ class CoreMixerTest(unittest.TestCase):
         self.mixer.get_mute.assert_called_once_with()
 
     def test_set_mute(self):
+        self.mixer.set_mute.return_value.get.return_value = True
         self.core.mixer.set_mute(True)
 
         self.mixer.set_mute.assert_called_once_with(True)
@@ -92,3 +94,63 @@ class CoreNoneMixerListenerTest(unittest.TestCase):
     def test_forwards_mixer_mute_changed_event_to_frontends(self, send):
         self.core.mixer.set_mute(mute=True)
         self.assertEqual(send.call_count, 0)
+
+
+class MockBackendCoreMixerBase(unittest.TestCase):
+
+    def setUp(self):  # noqa: N802
+        self.mixer = mock.Mock()
+        self.mixer.actor_ref.actor_class.__name__ = 'DummyMixer'
+        self.core = core.Core(mixer=self.mixer, backends=[])
+
+
+class GetVolumeBadBackendTest(MockBackendCoreMixerBase):
+
+    def test_backend_raises_exception(self):
+        self.mixer.get_volume.return_value.get.side_effect = Exception
+        self.assertEqual(self.core.mixer.get_volume(), None)
+
+    def test_backend_returns_too_small_value(self):
+        self.mixer.get_volume.return_value.get.return_value = -1
+        self.assertEqual(self.core.mixer.get_volume(), None)
+
+    def test_backend_returns_too_large_value(self):
+        self.mixer.get_volume.return_value.get.return_value = 1000
+        self.assertEqual(self.core.mixer.get_volume(), None)
+
+    def test_backend_returns_wrong_type(self):
+        self.mixer.get_volume.return_value.get.return_value = '12'
+        self.assertEqual(self.core.mixer.get_volume(), None)
+
+
+class SetVolumeBadBackendTest(MockBackendCoreMixerBase):
+
+    def test_backend_raises_exception(self):
+        self.mixer.set_volume.return_value.get.side_effect = Exception
+        self.assertFalse(self.core.mixer.set_volume(30))
+
+    def test_backend_returns_wrong_type(self):
+        self.mixer.set_volume.return_value.get.return_value = 'done'
+        self.assertFalse(self.core.mixer.set_volume(30))
+
+
+class GetMuteBadBackendTest(MockBackendCoreMixerBase):
+
+    def test_backend_raises_exception(self):
+        self.mixer.get_mute.return_value.get.side_effect = Exception
+        self.assertEqual(self.core.mixer.get_mute(), None)
+
+    def test_backend_returns_wrong_type(self):
+        self.mixer.get_mute.return_value.get.return_value = '12'
+        self.assertEqual(self.core.mixer.get_mute(), None)
+
+
+class SetMuteBadBackendTest(MockBackendCoreMixerBase):
+
+    def test_backend_raises_exception(self):
+        self.mixer.set_mute.return_value.get.side_effect = Exception
+        self.assertFalse(self.core.mixer.set_mute(True))
+
+    def test_backend_returns_wrong_type(self):
+        self.mixer.set_mute.return_value.get.return_value = 'done'
+        self.assertFalse(self.core.mixer.set_mute(True))